statistics of urls are now taken directly from the database instead of being incremented one by one during processing, this also caused inconsistencies with the numbers shown and numbers in the database.

+++ curation-module-core/src/main/java/eu/clarin/cmdi/curation/report/CMDInstanceReport.java
@@ -301,10 +301,10 @@ public class CMDInstanceReport implements Report<CollectionReport> {
     @XmlRootElement
     @XmlAccessorType(XmlAccessType.FIELD)
     public static class URLReport {
-        public int numOfLinks;
-        public int numOfUniqueLinks;
-        public int numOfCheckedLinks;
-        public int numOfBrokenLinks;
+        public long numOfLinks;
+        public long numOfUniqueLinks;
+        public long numOfCheckedLinks;
+        public long numOfBrokenLinks;
         public Double percOfValidLinks;
 
         public URLReport() {
